requirements.txt

# requirements.txt
streamlit
pandas
request
openai